apply feedback from PR review
